test: move command tests to domain tests
