+++ HaSpMan.sln
@@ -26,6 +26,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		.editorconfig = .editorconfig
 	EndProjectSection
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Domain.Test", "Domain.Test\Domain.Test.csproj", "{7282D5DD-7A7A-493A-BA08-C1F32A134238}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -144,6 +146,18 @@ Global
 		{09963342-F8AA-405E-90C7-1B3F0F66D8DD}.Release|x64.Build.0 = Release|Any CPU
 		{09963342-F8AA-405E-90C7-1B3F0F66D8DD}.Release|x86.ActiveCfg = Release|Any CPU
 		{09963342-F8AA-405E-90C7-1B3F0F66D8DD}.Release|x86.Build.0 = Release|Any CPU
+		{7282D5DD-7A7A-493A-BA08-C1F32A134238}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{7282D5DD-7A7A-493A-BA08-C1F32A134238}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7282D5DD-7A7A-493A-BA08-C1F32A134238}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{7282D5DD-7A7A-493A-BA08-C1F32A134238}.Debug|x64.Build.0 = Debug|Any CPU
+		{7282D5DD-7A7A-493A-BA08-C1F32A134238}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{7282D5DD-7A7A-493A-BA08-C1F32A134238}.Debug|x86.Build.0 = Debug|Any CPU
+		{7282D5DD-7A7A-493A-BA08-C1F32A134238}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{7282D5DD-7A7A-493A-BA08-C1F32A134238}.Release|Any CPU.Build.0 = Release|Any CPU
+		{7282D5DD-7A7A-493A-BA08-C1F32A134238}.Release|x64.ActiveCfg = Release|Any CPU
+		{7282D5DD-7A7A-493A-BA08-C1F32A134238}.Release|x64.Build.0 = Release|Any CPU
+		{7282D5DD-7A7A-493A-BA08-C1F32A134238}.Release|x86.ActiveCfg = Release|Any CPU
+		{7282D5DD-7A7A-493A-BA08-C1F32A134238}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
